server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
cryptography
mysql-connector-python
